Change parameters of watermark
